+++ src/vite.d.ts
@@ -1,16 +1,9 @@
-import type { Plugin as VitePlugin } from "vite";
+import type { Plugin } from "vite";
 
 export declare const downwind: (opts?: {
   /**
    * Used to reduce the number of scanned files.
-   * @default (id, code) =>
-   *   id.endsWith(".tsx") ||
-   *   (id.endsWith(".ts") && code.includes("@downwind-scan")),
+   * @default [{ id: /\.tsx$/ }, { id: /\.ts$/, code: /@downwind-scan/ }]
    */
-  shouldScan?: (id: string, code: string) => boolean;
-  /**
-   * Number of millisecond without new scan to wait before generating utils
-   * @default 200
-   */
-  buildIntervalCheckMs?: number;
-}) => VitePlugin[];
+  filters?: { id?: StringFilter; code?: StringFilter }[];
+}) => Plugin[];