+++ src/esbuild.d.ts
@@ -12,8 +12,10 @@ export declare const downwind: (opts?: {
    */
   shouldScan?: (path: string, code: string) => boolean;
   /**
-   * Number of millisecond without new scan to wait before generating utils
+   * Number of milliseconds without new scan to wait before generating utils
+   * `experimental-double-build` is an attempt to use a second in memeory build to scan all the files instead of using a timeout.
+   * This will consume more ressources but it avoids build randomlly failling on low-end environments like GH actions.
    * @default 50
    */
-  intervalCheckMs?: number;
+  idleMs?: number | "experimental-double-build";
 }) => ESBuildPlugin;